# Misc/CastleSwaggerApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo API",
    "version": "v1"
  },
  "paths": {
...
    "/commander/{name}": {
      "get": {
        "tags": [
          "Commander"
        ],
        "operationId": "GetCommander",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Commander"
        ],
        "operationId": "StopCommander",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/hostedfiles": {
      "get": {
        "tags": [
          "HostedFile"
        ],
        "operationId": "GetHostedFiles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostedFileResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostedFileResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostedFileResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "HostedFile"
        ],
        "operationId": "HostFile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostFileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/HostFileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/HostFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HostedFileResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedFileResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedFileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/hostedfiles/{id}": {
      "get": {
        "tags": [
          "HostedFile"
        ],
        "operationId": "GetHostedFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HostedFileResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedFileResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedFileResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "HostedFile"
        ],
        "operationId": "DeleteHostedFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "COMMANDER_TYPE": {
        "enum": [
          0
        ],
        "type": "integer",
        "format": "int32"
      },
      "Commander": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/COMMANDER_TYPE"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "bindPort": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "tls": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "HostFileRequest": {
        "type": "object",
        "properties": {
          "commander": {
            "type": "string",
            "nullable": true
          },
          "uri": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "bytes": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HostedFileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "commander": {
            "type": "string",
            "nullable": true
          },
          "uri": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}